aarch32: force 64-bit literals when needed

+++ src/arch/armv8/armv8-a/vm.c
@@ -13,7 +13,7 @@ void vcpu_arch_profile_init(struct vcpu* vcpu, struct vm* vm) {
 
     paddr_t root_pt_pa;
     mem_translate(&cpu()->as, (vaddr_t)vm->as.pt.root, &root_pt_pa);
-    sysreg_vttbr_el2_write(((vm->id << VTTBR_VMID_OFF) & VTTBR_VMID_MSK) |
+    sysreg_vttbr_el2_write((((uint64_t)vm->id << VTTBR_VMID_OFF) & VTTBR_VMID_MSK) |
                        (root_pt_pa & ~VTTBR_VMID_MSK));
 
     ISB();  // make sure vmid is commited befor tlbi
